requirements.txt
flask
rdkit-pypi
pandas
openpyxl
requests
